+++ tests/Mysql/InsertTest.php
@@ -313,4 +313,23 @@ class InsertTest extends Common\InsertTest
         $actual = $this->query->getBindValues();
         $this->assertSame($expect, $actual);
     }
+
+    public function testOnConflictNotSupported()
+    {
+        $this->expectException('Aura\SqlQuery\Exception\BadMethodCallException');
+        $this->expectExceptionMessage("doesn't support ON CONFLICT clause");
+        $this->query->onConflict('id');
+    }
+
+    public function testOrReplaceWithOnDuplicateKeyUpdateThrowsException()
+    {
+        $this->query->orReplace()
+                    ->into('t1')
+                    ->cols(array('c1'))
+                    ->onDuplicateKeyUpdate('c1', 'new-val');
+
+        $this->expectException('Aura\SqlQuery\Exception\LogicException');
+        $this->expectExceptionMessage('A REPLACE statement cannot take an ON DUPLICATE KEY UPDATE clause.');
+        $this->query->__toString();
+    }
 }